# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "User": {
        "properties": {
          "address": {
            "type": "string"
          },
          "emailOrPhone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "skills": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "linkedIn": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "tz": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "ROLE_USER"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "emailOrPhone"
        ],
        "example": {
          "id": "c166ff1b-bc2c-4656-b363-d121075b8f6a"
        }
      },
...
      "AuthEthLoginDto": {
        "properties": {
          "signature": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "signature",
          "address"
        ]
      },
      "AuthSolanaLoginDto": {
        "properties": {
          "signature": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "signature",
          "address"
        ]
      },
      "AuthNonceRequestDto": {
        "properties": {
          "address": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "address"
        ]
      },
      "AuthRefreshTokenDto": {
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "refreshToken"
        ]
      },
      "AuthTonProofDomainDto": {
        "properties": {
          "lengthBytes": {
            "type": "number"
          },
          "value": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "lengthBytes",
          "value"
        ]
      },
      "AuthTonProofPayloadDto": {
        "properties": {
          "timestamp": {
            "type": "number"
          },
          "domain": {
            "$ref": "#/components/schemas/AuthTonProofDomainDto"
          },
          "payload": {
            "minLength": 1,
            "type": "string"
          },
          "signature": {
            "minLength": 1,
            "type": "string"
          },
          "state_init": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "domain",
          "payload",
          "signature",
          "state_init"
        ]
      },
      "AuthTonLoginDto": {
        "properties": {
          "address": {
            "minLength": 1,
            "type": "string"
          },
          "network": {
            "minLength": 1,
            "type": "string"
          },
          "public_key": {
            "minLength": 1,
            "type": "string"
          },
          "proof": {
            "$ref": "#/components/schemas/AuthTonProofPayloadDto"
          }
        },
        "type": "object",
        "required": [
          "address",
          "network",
          "public_key",
          "proof"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (same value the API returns on login)"
      }
    }
  },
  "info": {
    "title": "API schema",
    "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/eth": {
      "post": {
        "operationId": "AuthController.loginEth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthEthLoginDto"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Authorization": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              },
              "Refresh-Token": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login with Ethereum wallet",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/solana": {
      "post": {
        "operationId": "AuthController.loginSolana",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthSolanaLoginDto"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Authorization": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              },
              "Refresh-Token": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login with Solana wallet",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/ton": {
      "post": {
        "operationId": "AuthController.checkProofHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthTonLoginDto"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Authorization": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              },
              "Refresh-Token": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login with TON Connect proof",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/nonce": {
      "post": {
        "operationId": "AuthController.nonce",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthNonceRequestDto"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Request nonce for Ethereum login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController.refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshTokenDto"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Authorization": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              },
              "Refresh-Token": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "JWT token rotation",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ]
      }
    },
    "/api/auth/status": {
      "get": {
        "operationId": "AuthController.status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Current user from JWT",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}